Quiz_demo/requirements..txt
load_dotenv==0.1.0
python-docx==1.2.0
pdfplumber==0.11.7
google-genai
pypandoc==1.15
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
